+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Link unacked PS update stream to PS module on product definitions sync (OSIDB-629)
+- Increase PS component name length from 100 to 255 characters (OSIDB-635)
 
 ## [2.3.2] - 2022-11-28
 ### Changed